pyproject.toml
[project]
name = "brother_print"
version = "0.1.0"
dependencies = ["pycups"]
